+++ app/ids-modal/nested.js
@@ -1,20 +1,20 @@
-import IdsModal from '../../src/ids-modal/ids-modal';
+import IdsModal, { IdsModalButton } from '../../src/ids-modal';
 import IdsButton from '../../src/ids-button/ids-button';
 
 document.addEventListener('DOMContentLoaded', () => {
   const parentTriggerId = '#parent-modal-trigger-btn';
   const parentTriggerBtn = document.querySelector(parentTriggerId);
   const parentModal = document.querySelector('#parent-modal');
-  const parentModalCloseBtn = parentModal.querySelector('ids-button');
 
   const nestedTriggerId = '#nested-modal-trigger-btn';
   const nestedTriggerBtn = document.querySelector(nestedTriggerId);
   const nestedModal = document.querySelector('#nested-modal');
-  const nestedModalCloseBtn = nestedModal.querySelector('ids-button');
 
   // Links the Modals to their trigger buttons
   parentModal.target = parentTriggerBtn;
+  parentModal.trigger = 'click';
   nestedModal.target = nestedTriggerBtn;
+  nestedModal.trigger = 'click';
 
   // Disable the trigger buttons when showing their Modals.
   parentModal.addEventListener('beforeshow', () => {
@@ -27,12 +27,14 @@ document.addEventListener('DOMContentLoaded', () => {
   });
 
   // Close the modals when their "close" buttons are clicked
-  parentModalCloseBtn.addEventListener('click', () => {
-    parentModal.hide();
-  });
-  nestedModalCloseBtn.addEventListener('click', () => {
+  parentModal.onButtonClick = (buttonEl) => {
+    if (buttonEl.cancel) {
+      parentModal.hide();
+    }
+  };
+  nestedModal.onButtonClick = () => {
     nestedModal.hide();
-  });
+  };
 
   // When the modals are fully hidden, re-enable their trigger buttons
   parentModal.addEventListener('hide', () => {