+++ src/ids-container/ids-container.js
@@ -10,7 +10,8 @@ import {
 // Import Mixins
 import {
   IdsEventsMixin,
-  IdsThemeMixin
+  IdsThemeMixin,
+  IdsLocaleMixin
 } from '../ids-mixins';
 
 import styles from './ids-container.scss';
@@ -21,11 +22,12 @@ import styles from './ids-container.scss';
  * @inherits IdsElement
  * @mixes IdsThemeMixin
  * @mixes IdsEventsMixin
+ * @mixes IdsLocaleMixin
  * @part container - the entire container element
  */
 @customElement('ids-container')
 @scss(styles)
-class IdsContainer extends mix(IdsElement).with(IdsEventsMixin, IdsThemeMixin) {
+class IdsContainer extends mix(IdsElement).with(IdsEventsMixin, IdsThemeMixin, IdsLocaleMixin) {
   constructor() {
     super();
   }
@@ -35,16 +37,24 @@ class IdsContainer extends mix(IdsElement).with(IdsEventsMixin, IdsThemeMixin) {
    */
   connectedCallback() {
     super.connectedCallback();
+
+    /* istanbul ignore next */
+    if (this.reset) {
+      this.#addReset();
+    }
   }
 
   /**
-   * Return the properties we handle as getters/setters
-   * @returns {Array} The properties in an array
+   * Return the attributes we handle as getters/setters
+   * @returns {Array} The attributes in an array
    */
   static get attributes() {
     return [
+      attributes.LANGUAGE,
+      attributes.LOCALE,
       attributes.MODE,
       attributes.PADDING,
+      attributes.RESET,
       attributes.SCROLLABLE,
       attributes.VERSION
     ];
@@ -63,15 +73,12 @@ class IdsContainer extends mix(IdsElement).with(IdsEventsMixin, IdsThemeMixin) {
    * @param {string} value sets the padding to the container
    */
   set padding(value) {
-    const double = Number(value) * 2;
     this.container.style.padding = `${value}px`;
-    this.container.style.height = `calc(100% - ${double}px)`;
-    this.container.style.width = `calc(100% - ${double}px)`;
-    this.setAttribute('padding', value.toString());
+    this.setAttribute(attributes.PADDING, value.toString());
   }
 
   get padding() {
-    return this.getAttribute('padding');
+    return this.getAttribute(attributes.PADDING);
   }
 
   /**
@@ -80,18 +87,41 @@ class IdsContainer extends mix(IdsElement).with(IdsEventsMixin, IdsThemeMixin) {
    */
   set scrollable(value) {
     if (stringUtils.stringToBool(value)) {
-      this.setAttribute('scrollable', 'true');
-      this.container.setAttribute('scrollable', 'true');
+      this.setAttribute(attributes.SCROLLABLE, 'true');
+      this.container.setAttribute(attributes.SCROLLABLE, 'true');
       this.container.setAttribute('tabindex', '0');
       return;
     }
 
-    this.setAttribute('scrollable', 'false');
-    this.container.setAttribute('scrollable', 'false');
+    this.setAttribute(attributes.SCROLLABLE, 'false');
+    this.container.setAttribute(attributes.SCROLLABLE, 'false');
     this.container.removeAttribute('tabindex');
   }
 
-  get scrollable() { return this.getAttribute('scrollable') || 'true'; }
+  get scrollable() { return this.getAttribute(attributes.SCROLLABLE) || 'true'; }
+
+  /**
+   * Add the reset to the body
+   * @private
+   */
+  #addReset() {
+    document.querySelector('body').style.margin = '0';
+  }
+
+  /**
+   * If set to true body element will get reset
+   * @param {boolean|string} value true of false
+   */
+  set reset(value) {
+    if (stringUtils.stringToBool(value)) {
+      this.#addReset();
+      return;
+    }
+    this.removeAttribute(attributes.RESET);
+    document.querySelector('body').style.margin = '';
+  }
+
+  get reset() { return this.getAttribute(attributes.RESET) || 'true'; }
 }
 
 export default IdsContainer;