+++ src/ids-header/ids-header.d.ts
@@ -1,0 +1,7 @@
+export default class IdsHeader extends HTMLElement {
+  /** Set the theme version */
+  version: 'new' | 'classic' | string;
+
+  /** Set the color */
+  color: '#0072ed' | string;
+}