+++ src/ids-icon/ids-icon.js
@@ -4,8 +4,15 @@ import {
   scss,
   customElement,
   attributes,
-  stringUtils
+  stringUtils,
+  mix
 } from '../ids-base';
+
+// Import Mixins
+import {
+  IdsEventsMixin,
+  IdsLocaleMixin
+} from '../ids-mixins';
 
 import styles from './ids-icon.scss';
 
@@ -21,30 +28,67 @@ const sizes = {
  * IDS Icon Component
  * @type {IdsIcon}
  * @inherits IdsElement
+ * @mixes IdsLocaleMixin
  */
 @customElement('ids-icon')
 @scss(styles)
-class IdsIcon extends IdsElement {
+class IdsIcon extends mix(IdsElement).with(IdsEventsMixin, IdsLocaleMixin) {
   constructor() {
     super();
   }
 
+  connectedCallback() {
+    super.connectedCallback();
+    this.handleEvents();
+  }
+
   /**
-   * Return the properties we handle as getters/setters
-   * @returns {Array} The properties in an array
+   * Return the attributes we handle as getters/setters
+   * @returns {Array} The attributes in an array
    */
   static get attributes() {
-    return [attributes.ICON, attributes.SIZE, attributes.VERTICAL];
+    return [
+      ...super.attributes,
+      attributes.LANGUAGE,
+      attributes.LOCALE,
+      attributes.ICON,
+      attributes.SIZE,
+      attributes.VERTICAL
+    ];
+  }
+
+  /**
+   * Handle change events
+   */
+  handleEvents() {
+    this.offEvent('languagechange.container');
+    this.onEvent('languagechange.container', this.closest('ids-container'), async (e) => {
+      await this.setLanguage(e.detail.language.name);
+      if (this.isFlipped(this.icon)) {
+        this.container.classList.add('flipped');
+      } else {
+        this.container.classList.remove('flipped');
+      }
+    });
+
+    this.offEvent('languagechange.icon');
+    this.onEvent('languagechange.icon', this, async (e) => {
+      await this.locale.setLanguage(e.detail.language.name);
+      if (this.isFlipped(this.icon)) {
+        this.container.classList.add('flipped');
+      } else {
+        this.container.classList.remove('flipped');
+      }
+    });
   }
 
   /**
    * Create the Template for the contents
-   *
    * @returns {string} The template
    */
   template() {
     const size = sizes[this.size];
-    return `<svg xmlns="http://www.w3.org/2000/svg" stroke="currentColor" fill="none" height="${size}" width="${size}" viewBox="0 0 18 18" focusable="false" aria-hidden="true" role="presentation">
+    return `<svg xmlns="http://www.w3.org/2000/svg"${this.isFlipped(this.icon) ? ` class="flipped"` : ''} stroke="currentColor" fill="none" height="${size}" width="${size}" viewBox="0 0 18 18" aria-hidden="true">
       ${this.iconData()}
     </svg>`;
   }
@@ -55,6 +99,118 @@ class IdsIcon extends IdsElement {
    */
   iconData() {
     return pathData[this.icon];
+  }
+
+  /**
+   * Some icons are flipped in RTL Mode
+   * @param {string} iconName icon name to check
+   * @returns {boolean} true if flipped / rtl
+   */
+  isFlipped(iconName) {
+    const flippedIcons = [
+      'add-grid-record',
+      'add-grid-row',
+      'attach',
+      'bullet-list',
+      'bullet-steps',
+      'caret-left',
+      'caret-right',
+      'cart',
+      'cascade',
+      'change-font',
+      'clear-screen',
+      'clockwise-90',
+      'close-cancel',
+      'close-save',
+      'closed-folder',
+      'collapse-app-tray',
+      'contacts',
+      'copy-from',
+      'copy-mail',
+      'copy-url',
+      'counter-clockwise-90',
+      'create-report',
+      'cut',
+      'delete-grid-record',
+      'delete-grid-row',
+      'display',
+      'document',
+      'drilldown',
+      'duplicate',
+      'employee-directory',
+      'expand-app-tray',
+      'export',
+      'export-2',
+      'export-to-pdf',
+      'first-page',
+      'folder',
+      'generate-key',
+      'get-more-rows',
+      'group-selection',
+      'headphones',
+      'help',
+      'helper-list-select',
+      'history',
+      'import',
+      'invoice-released',
+      'key',
+      'language',
+      'last-page',
+      'launch',
+      'left-align',
+      'left-arrow',
+      'left-text-align',
+      'logout',
+      'new-document',
+      'new-expense-report',
+      'new-time-sheet',
+      'new-travel-plan',
+      'next-page',
+      'no-attachment',
+      'no-comment',
+      'no-filter',
+      'paste',
+      'phone',
+      'previous-page',
+      'queries',
+      'quick-access',
+      'redo',
+      'refresh',
+      'refresh-current',
+      'restore-user',
+      'right-align',
+      'right-arrow',
+      'right-text-align',
+      'run-quick-access',
+      'save',
+      'save-close',
+      'save-new',
+      'search',
+      'search-folder',
+      'search-list',
+      'search-results-history',
+      'select',
+      'send',
+      'send-submit',
+      'show-last-x-days',
+      'special-item',
+      'stacked',
+      'tack',
+      'timesheet',
+      'tree-collapse',
+      'tree-expand',
+      'undo',
+      'unsubscribe',
+      'update-preview',
+      'zoom-100',
+      'zoom-in',
+      'zoom-out'
+    ];
+
+    if (this.locale.isRTL() && flippedIcons.includes(iconName)) {
+      return true;
+    }
+    return false;
   }
 
   /**