+++ src/ids-locale/cultures/el-GR.js
@@ -1,0 +1,64 @@
+// Info Comes from http://www.unicode.org/Public/cldr/25/
+const locale = {
+  // layout/language
+  language: 'el',
+  englishName: 'Greek (Greece)',
+  nativeName: 'Ελληνικά (Ελλάδα)',
+  // layout/orientation/@characters
+  direction: 'left-to-right',
+  // ca-gregorian
+  calendars: [{
+    name: 'gregorian',
+    // ca-gregorian/main/dates/calendars/gregorian/dateFormats/
+    dateFormat: {
+      separator: '/', // Infered
+      timeSeparator: ':',
+      short: 'd/M/yyyy', // use four digit year
+      medium: 'd MMM yyyy',
+      long: 'd MMMM yyyy',
+      full: 'EEEE, d MMMM yyyy',
+      month: 'd MMMM',
+      year: 'MMMM yyyy',
+      dayOfWeek: 'EEE d',
+      timestamp: 'h:mm:ss a',
+      timestampMillis: 'h:mm:ss.SSS a',
+      hour: 'h:mm a',
+      datetime: 'd/M/yyyy h:mm a',
+      datetimeMillis: 'd/M/yyyy h:mm:ss.SSS a',
+      timezone: 'd/M/yyyy h:mm a zz',
+      timezoneLong: 'd/M/yyyy h:mm a zzzz'
+    }, // Infered short + short gregorian/dateTimeFormats
+    // ca-gregorian/main/dates/calendars/gregorian/days/format/short or abbreviated (2 digit)
+    days: {
+      wide: ['Κυριακή', 'Δευτέρα', 'Τρίτη', 'Τετάρτη', 'Πέμπτη', 'Παρασκευή', 'Σάββατο'],
+      abbreviated: ['Κυρ', 'Δευ', 'Τρί', 'Τετ', 'Πέμ', 'Παρ', 'Σάβ'],
+      narrow: ['Κ', 'Δ', 'Τ', 'Τ', 'Π', 'Π', 'Σ']
+    },
+    // ca-gregorian/main/dates/calendars/gregorian/months/format/wide
+    months: {
+      wide: ['Ιανουάριος', 'Φεβρουάριος', 'Μάρτιος', 'Απρίλιος', 'Μάιος', 'Ιούνιος', 'Ιούλιος', 'Αύγουστος', 'Σεπτέμβριος', 'Οκτώβριος', 'Νοέμβριος', 'Δεκέμβριος'],
+      abbreviated: ['Ιαν', 'Φεβ', 'Μαρ', 'Απρ', 'Μαΐ', 'Ιουν', 'Ιουλ', 'Αυγ', 'Σεπ', 'Οκτ', 'Νοε', 'Δεκ']
+    },
+    // ca-gregorian/main/dates/calendars/gregorian/timeFormats/short
+    timeFormat: 'h:mm a',
+    // ca-gregorian/main/dates/calendars/gregorian/dayPeriods/wide
+    dayPeriods: ['π.μ.', 'μ.μ.'],
+    firstDayofWeek: 1 // Starts on Mon
+  }],
+  // numbers/currencyFormats-numberSystem-latn/standard
+  currencySign: '€',
+  currencyFormat: '### ¤',
+  // numbers/symbols-numberSystem-latn
+  numbers: {
+    percentSign: '%',
+    percentFormat: '### %',
+    percentSuffix: ' %',
+    percentPrefix: '',
+    minusSign: '-',
+    decimal: ',',
+    group: '.',
+    groupSizes: [3, 3]
+  }
+};
+
+export { locale };