+++ src/ids-locale/cultures/th-TH.js
@@ -1,0 +1,64 @@
+// Info Comes from http://www.unicode.org/Public/cldr/25/
+const locale = {
+  // layout/language
+  language: 'th',
+  englishName: 'Thai (Thailand)',
+  nativeName: 'ไทย (ไทย)',
+  // layout/orientation/@characters
+  direction: 'left-to-right',
+  // ca-gregorian
+  calendars: [{
+    name: 'gregorian',
+    // ca-gregorian/main/dates/calendars/gregorian/dateFormats/
+    dateFormat: {
+      separator: '/', // Infered
+      timeSeparator: ':',
+      short: 'd/M/yyyy', // use four digit year
+      medium: 'd MMM yyyy',
+      long: 'd MMMM yyyy',
+      full: 'EEEEที่ d MMMM G yyyy',
+      month: 'd MMMM',
+      year: 'MMMM yyyy',
+      dayOfWeek: 'EEE d',
+      timestamp: 'HH:mm:ss',
+      timestampMillis: 'HH:mm:ss.SSS',
+      hour: 'HH:mm',
+      datetime: 'd/M/yyyy HH:mm',
+      datetimeMillis: 'd/M/yyyy HH:mm:ss.SSS',
+      timezone: 'd/M/yyyy HH:mmm zz',
+      timezoneLong: 'd/M/yyyy HH:mm zzzz'
+    }, // Infered short + short gregorian/dateTimeFormats
+    // ca-gregorian/main/dates/calendars/gregorian/days/format/short or abbreviated (2 digit)
+    days: {
+      wide: ['วันอาทิตย์', 'วันจันทร์', 'วันอังคาร', 'วันพุธ', 'วันพฤหัสบดี', 'วันศุกร์', 'วันเสาร์'],
+      abbreviated: ['อา.', 'จ.', 'อ.', 'พ.', 'พฤ.', 'ศ.', 'ส.'],
+      narrow: ['อา', 'จ', 'อ', 'พ', 'พฤ', 'ศ', 'ส']
+    },
+    // ca-gregorian/main/dates/calendars/gregorian/months/format/wide
+    months: {
+      wide: ['มกราคม', 'กุมภาพันธ์', 'มีนาคม', 'เมษายน', 'พฤษภาคม', 'มิถุนายน', 'กรกฎาคม', 'สิงหาคม', 'กันยายน', 'ตุลาคม', 'พฤศจิกายน', 'ธันวาคม'],
+      abbreviated: ['ม.ค.', 'ก.พ.', 'มี.ค.', 'เม.ย.', 'พ.ค.', 'มิ.ย.', 'ก.ค.', 'ส.ค.', 'ก.ย.', 'ต.ค.', 'พ.ย.', 'ธ.ค.']
+    },
+    // ca-gregorian/main/dates/calendars/gregorian/timeFormats/short
+    timeFormat: 'HH:mm',
+    // ca-gregorian/main/dates/calendars/gregorian/dayPeriods/wide
+    dayPeriods: ['ก่อนเที่ยง', 'หลังเที่ยง'],
+    firstDayofWeek: 1 // Starts on Monday
+  }],
+  // numbers/currencyFormats-numberSystem-latn/standard
+  currencySign: 'kr',
+  currencyFormat: '¤###',
+  // numbers/symbols-numberSystem-latn
+  numbers: {
+    percentSign: '%',
+    percentFormat: '### %',
+    percentSuffix: ' %',
+    percentPrefix: '',
+    minusSign: '-',
+    decimal: '.',
+    group: ',',
+    groupSizes: [3, 3]
+  }
+};
+
+export { locale };