+++ src/ids-mixins/ids-theme-mixin.js
@@ -20,7 +20,6 @@ const IdsThemeMixin = (superclass) => class extends superclass {
 
   connectedCallback() {
     super.connectedCallback?.();
-
     this.initThemeHandlers();
   }