+++ src/ids-trigger-field/ids-trigger-button.js
@@ -26,8 +26,8 @@ class IdsTriggerButton extends IdsButton {
   }
 
   /**
-   * Return the properties we handle as getters/setters
-   * @returns {Array} The properties in an array
+   * Return the attributes we handle as getters/setters
+   * @returns {Array} The attributes in an array
    */
   static get attributes() {
     return [attributes.CSS_CLASS,