+++ test/ids-notification-banner/ids-notification-banner-e2e-test.js
@@ -1,0 +1,14 @@
+describe('Ids Notification Banner e2e Tests', () => {
+  const exampleUrl = 'http://localhost:4444/ids-notification-banner';
+
+  it('should not have errors', async () => {
+    await page.goto(exampleUrl, { waitUntil: ['domcontentloaded', 'networkidle0'] });
+    await expect(page.title()).resolves.toMatch('IDS Notification Banner Component');
+  });
+
+  it('should pass Axe accessibility tests', async () => {
+    await page.setBypassCSP(true);
+    await page.goto(exampleUrl, { waitUntil: ['networkidle2', 'load'] });
+    await expect(page).toPassAxeTests();
+  });
+});